move e2e and configuration tests to more appropriate spots
