+++ deeplearning4j-modelimport/src/main/java/org/deeplearning4j/nn/modelimport/keras/KerasModel.java
@@ -68,7 +68,7 @@ public class KerasModel {
     int truncatedBPTT = 0; // truncated BPTT value
     int kerasMajorVersion;
 
-    protected KerasModel() {
+    public KerasModel() {
     }
 
     public KerasModelBuilder modelBuilder() {
